add Session.key property for django 1.3
sample couchdb document in the database after django.contrib.auth.login:

{
   "_id": "2a7dfe71e7e44303d72dc2212ffae481",
   "_rev": "2-d6e2d40de4e7970e4e820781908961aa",
   "doc_type": "Session",
   "key": "bf7e58c572f6437be04804d3c3626c71",
   "expire_date": "2011-06-28T13:42:09Z",
   "session_key": null,
   "session_data": "NTU5Mzk3ZjZjZWVjMmRkNmVmNjNhYTFlOWMwMTRjMjJmODNiOTg2NDqAAn1xAShVEl9hdXRoX3Vz\nZXJfYmFja2VuZHECWCwAAABkamFuZ29fY291Y2hkYl91dGlscy5hdXRoLkNvdWNoREJBdXRoQmFj\na2VuZHEDVQ1fYXV0aF91c2VyX2lkcQRYBAAAAG1pa2VxBXUu\n"
}

Not sure what key and session_key are meant for, since the latter is
stored in the document via:
session.key = self.session_key
in SessionStore.save, but the field is not in the model and thus not
available for writing when the document is fetched.

Possible solution?
session.session_key = self.session_key